feat(profile): Settings as a tab, unified banner upload, sidebar headers
- Settings: extracted the two-pane scroll-spy layout into a shared
  <SettingsPanel> component; rendered both at /settings (preserved as
  a deep-link target) and as a real ?tab=settings panel on the user's
  own profile. The top-bar Settings tab now uses the normal tab-href
  logic (no more /settings shortcut), so the URL becomes
  /profile/<handle>?tab=settings consistently.
- Banner upload: refactored to mirror the avatar's always-visible
  Camera button instead of hover-reveal — both affordances now share
  the same interaction model. Empty banner renders a grey gradient
  box with the same "Change banner" pill so the upload path is
  identical with or without an image.
- Sidebar edit: added "Main website" / "Additional links" sub-headers
  above the corresponding inputs so each block has a clear label.
- Removed the Groups stat from the profile-overview stat row.
