+++ src/components/profile/banner-upload.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { useState, useRef } from "react";
-import { Image as ImageIcon } from "lucide-react";
+import { Camera } from "lucide-react";
 import LoadingSpinner from "@/components/ui/loading-spinner";
 
 export interface BannerUploadProps {
@@ -10,6 +10,18 @@ export interface BannerUploadProps {
   isUploading: boolean;
 }
 
+const MAX_FILE_SIZE = 4 * 1024 * 1024; // 4MB — Vercel serverless limit is ~4.5MB
+const ALLOWED_TYPES = ["image/jpeg", "image/png", "image/webp"];
+
+/**
+ * Banner upload control. Mirrors `AvatarEditOverlay` in the sidebar:
+ * always renders a floating "Change banner" button in the corner so
+ * the affordance is the same regardless of whether a banner image is
+ * present or not. The button itself opens the file picker; the rest
+ * of the banner box is decorative (no whole-area click target, which
+ * matches the avatar behaviour and avoids the rect-click vs.
+ * pill-click inconsistency the user flagged).
+ */
 const BannerUpload: React.FC<BannerUploadProps> = ({
   currentBannerUrl,
   onUpload,
@@ -17,12 +29,8 @@ const BannerUpload: React.FC<BannerUploadProps> = ({
 }) => {
   const [previewUrl, setPreviewUrl] = useState<string | null>(null);
   const [error, setError] = useState<string | null>(null);
-  const [isHovered, setIsHovered] = useState(false);
   const [imgFailed, setImgFailed] = useState(false);
   const fileInputRef = useRef<HTMLInputElement>(null);
-
-  const MAX_FILE_SIZE = 4 * 1024 * 1024; // 4MB — Vercel serverless limit is ~4.5MB
-  const ALLOWED_TYPES = ["image/jpeg", "image/png", "image/webp"];
 
   const handleClick = () => {
     fileInputRef.current?.click();
@@ -34,38 +42,31 @@ const BannerUpload: React.FC<BannerUploadProps> = ({
 
     setError(null);
 
-    // Validate file type
     if (!ALLOWED_TYPES.includes(file.type)) {
       setError("Please select a JPEG, PNG, or WebP image");
       return;
     }
 
-    // Validate file size
     if (file.size > MAX_FILE_SIZE) {
       setError("Image must be 4MB or smaller");
       return;
     }
 
-    // Create preview
     const objectUrl = URL.createObjectURL(file);
     setPreviewUrl(objectUrl);
 
-    // Upload
     try {
       await onUpload(file);
     } catch (err) {
       console.error("Upload failed:", err);
       setError(err instanceof Error ? err.message : "Upload failed");
-      // Clear preview on error
       URL.revokeObjectURL(objectUrl);
       setPreviewUrl(null);
     }
 
-    // Reset file input
     e.target.value = "";
   };
 
-  // Clean up preview URL on unmount
   React.useEffect(() => {
     return () => {
       if (previewUrl) {
@@ -75,65 +76,63 @@ const BannerUpload: React.FC<BannerUploadProps> = ({
   }, [previewUrl]);
 
   const displayUrl = previewUrl || currentBannerUrl;
+  const hasImage = !!displayUrl && !imgFailed;
+  const hasPending = !!previewUrl;
 
-  // Reset the error flag when the displayed URL changes so users can
-  // recover after a failed load by uploading a new file.
   React.useEffect(() => {
     setImgFailed(false);
   }, [displayUrl]);
 
   return (
-    <div className="flex flex-col gap-2">
+    <div className="profile-banner-upload">
       <div
-        className="profile-card__banner relative cursor-pointer"
-        style={{ marginBottom: 0 }}
-        role="button"
-        tabIndex={0}
-        aria-label="Upload banner"
-        onClick={handleClick}
-        onKeyDown={(e) => { if (e.key === "Enter" || e.key === " ") { e.preventDefault(); handleClick(); } }}
-        onMouseEnter={() => setIsHovered(true)}
-        onMouseLeave={() => setIsHovered(false)}
+        className={
+          "profile-banner-upload__box" +
+          (hasImage ? "" : " profile-banner-upload__box--empty")
+        }
       >
-        {/* Banner image or placeholder (falls through to the gradient
-            background when the URL is missing or fails to load, instead
-            of showing the browser's broken-image icon). */}
-        {displayUrl && !imgFailed ? (
+        {hasImage ? (
+          // eslint-disable-next-line @next/next/no-img-element
           <img
-            src={displayUrl}
+            src={displayUrl!}
             alt="Profile banner"
+            className="profile-banner-upload__img"
             onError={() => setImgFailed(true)}
           />
         ) : null}
 
-        {/* Overlay */}
-        {(isHovered || isUploading) && (
-          <div className="absolute inset-0 bg-black bg-opacity-50 flex items-center justify-center gap-2">
-            {isUploading ? (
-              <LoadingSpinner size="sm" />
-            ) : (
-              <>
-                <ImageIcon className="h-6 w-6 text-white" />
-                <span className="text-white text-xs">Change banner</span>
-              </>
-            )}
-          </div>
-        )}
+        <button
+          type="button"
+          className="profile-banner-upload__btn"
+          onClick={handleClick}
+          aria-label={isUploading ? "Uploading banner" : "Change banner"}
+          title={isUploading ? "Uploading…" : "Change banner"}
+          disabled={isUploading}
+        >
+          {isUploading ? (
+            <LoadingSpinner size="sm" />
+          ) : (
+            <>
+              <Camera size={16} strokeWidth={1.75} aria-hidden />
+              <span className="profile-banner-upload__btn-label">
+                {hasPending ? "Replace banner" : "Change banner"}
+              </span>
+            </>
+          )}
+        </button>
 
-        {/* Hidden file input */}
         <input
           ref={fileInputRef}
           type="file"
           accept="image/jpeg,image/png,image/webp"
           onChange={handleFileChange}
-          className="hidden"
+          className="profile-banner-upload__input"
           disabled={isUploading}
         />
       </div>
 
-      {/* Error message */}
       {error && (
-        <p role="alert" className="text-caption text-[var(--color-error)]">
+        <p role="alert" className="profile-banner-upload__error">
           {error}
         </p>
       )}
